Let current task finish before shutdown
The shutdown event blocks all incoming requests, therefore we cannot wait for a get request for the current task to be received first before shutting down the app. So, the alternative for now is to at least let the current task finish, then program something in the Airflow triggerer that checks in S3 whether a specific program has been transcribed already or not.